CpG Neighborhoods from Genomic Distance

Clustering of CpG sites to define CpG neighborhoods based on distance (bp). INTERNAL USE ONLY.

cluster_sites(methytmle, window_size = 1000)

Arguments

methytmle
Object of class methytmle produced from an object of class GenomicRatioSet, but containing extra slots.

window_size
Numeric giving the number of base pairs used to define neighborhoods along the genome. CpG sites within this distance (bp) of one another are denoted as neighbors. Chromosome boundaries and other biological constraints are respected (see the documentation available for bumphunter::clusterMaker for details).

Value

An object of class methytmle with the "clusters" slot filled in. The "clusters" slot contains a numeric vector as long as the number of CpG sites. Each entry in the vector is a neighborhood assignment used in the estimation procedure.