FDR-MSA correction

Modified FDR Controlling Procedure for Multi-Stage Analyses (MJ van der Laan and C Tuglus, 2009, <doi:10.2202/1544-6115.1397>)

fdr_msa(pvals, total_obs)

Arguments

pvals
Numeric vector containing the p-values that result from any chosen statistical hypothesis testing procedure.

total_obs
Numeric indicating the total number of observations that would have been available for testing prior to the selection procedure employed in the multi-stage analysis performed.

Value

A numeric vector of corrected p-values, controlling the False Discovery Rate, using the method of Tuglus and van der Laan.

Examples

g <- 1e4
n <- 1e2
p <- abs(rnorm(n, mean = 1e-8, sd = 1e-2))
# treating the vector p as one of p-values, FDR-MSA may be applied
fdr_p <- fdr_msa(pvals = p, total_obs = g)