Differential Methylation with Classical Target Parameters

Computes the Targeted Minimum Loss-Based Estimate of the Average Treatment Effect (ATE) or the Risk Ratio, treating DNA methylation as an outcome (Y) and the indicated variable of interest (which ought to be binarized) as a treatment/exposure (A), using the neighbors of a given CpG site as the adjustment set (W). INTERNAL USE ONLY.

methyvim_tmle(target_site, methytmle_screened, var_of_interest,
  type = c("Beta", "Mval"), corr, obs_per_covar,
  target_param = c("ate", "rr"), g_lib = c("SL.mean", "SL.glm",
  "SL.glm.interaction"), Q_lib = c("SL.mean", "SL.glm", "SL.gam",
  "SL.earth"), family = c("gaussian", "binomial"), return_ic = FALSE)

Arguments

target_site
Numeric indicating the column containing the screened CpG site indices that will be looped over in TMLE procedure.

methytmle_screened
An object of class methytmle with clustered sites based on genomic windows containing the screened CpG site indices column.

var_of_interest
Numeric indicating the column index of the binarized variable of interest, treated as an exposure.

type
Character indicating the particular measure of DNA methylation to be used as the observed data in the estimation procedure, either Beta values or M-values. The data are accessed via minfi::getBeta or minfi::getM.

corr
Numeric indicating the maximum correlation that a neighboring site can have with the target site.

obs_per_covar
Numeric indicating the number of observations needed for for covariate included in W for downstream analysis. This ensures the data is sufficient to control for the covariates.

target_param
Character indicating the target causal parameter for which an estimator will be constructed and computed via targeted minimum loss-based estimation. Currently, this is limited to the Average Treatment Effect (ATE) and the Risk Ratio (RR), with routines from the tmle package being used for the computation.

g_lib
Character or vector of characters indicating the algorithms to be implemented in SuperLearner if tmle_type is set to "glm".

Q_lib
Character or vector of characters indicating the algorithms to be implemented in SuperLearner if tmle_type is set to "sl".

family
Character indicating the distribution to be implemented to describe the error distribution for regressions, generally "gaussian" for a continuous outcome and "binomial" for a binary outcome.

return_ic
Logical indicating whether an influence curve estimate should be returned for each site that passed through the filter.

Value

An data.frame containing the results of the Targeted Minimum Loss-based Estimation (TMLE) procedure for the target parameter of interest for a single CpG site, computed via the tmle package.